[UPD] base_wua_infrastructure: grant GIS privileges for drainage ditch, pressure sensor, and tank.

+++ base_wua_remotecontrol_rest/models/wua_waterconnection.py
@@ -28,7 +28,7 @@ class WuaWaterconnection(models.Model):
         [],
         string='Type of telecontrol associated')
 
-    conversion_factor = fields.Integer(
+    conversion_factor = fields.Float(
         string="Conversion Factor",
         required=True,
         default=1)
